Floats passés en Doubles et correction hauteur min max

+++ Synthese/mathutils.cpp
@@ -28,7 +28,7 @@ double MathUtils::interpolate(double before_p0, double p0, double p1, double aft
     return (a3 * t*t*t) + (a2 * t*t) + (a1 * t) + a0;
 }
 
-float MathUtils::ridge(float n, float s) {
+double MathUtils::ridge(double n, double s) {
     //std::cout<<s<<std::endl;
     return n>s?2*s-n:n;
 }
