+++ Synthese/ray.cpp
@@ -23,11 +23,11 @@ bool Ray::intersects(const Vector3D &a, const Vector3D &b, const Vector3D &c, Ve
     Vector3D w =u^v;//Vector3D::crossProduct(u, v);
 
     // Produit scalaire, si 0 pas de solution ou une infinité
-    float scalaire = direction*w;//Vector3D::dotProduct(direction, w); //direction.x()*w.x()+direction.y()*w.y()+direction.z()*w.z();
+    double scalaire = direction*w;//Vector3D::dotProduct(direction, w); //direction.x()*w.x()+direction.y()*w.y()+direction.z()*w.z();
     if(abs(scalaire) < 0.0001)
         return false;
-    float delta = a*w;//Vector3D::dotProduct(a, w);
-    float t = (delta - origine*w)/scalaire; //Vector3D::dotProduct(origine, w)
+    double delta = a*w;//Vector3D::dotProduct(a, w);
+    double t = (delta - origine*w)/scalaire; //Vector3D::dotProduct(origine, w)
 
     out = getPoint(t);
     return true;
@@ -209,11 +209,11 @@ bool Ray::intersectRayMarching(Terrain& t, const Vector3D &aBox, const Vector3D 
         //std::cout<<"return false"<<std::endl;
         return false;
     }
-    float distInOut = in.distanceToPoint(out);
+    double distInOut = in.distanceToPoint(out);
     Vector3D dir=direction;
     Ray r(in,dir);
     //std::cout << pas << " " << dir.x() << "," << dir.y() << "," << dir.z() << " / " << origine.x() << "," << origine.y() << "," << origine.z() << std::endl;
-    for(float tt = 0; tt <= distInOut+pas; tt+= pas)
+    for(double tt = 0; tt <= distInOut+pas; tt+= pas)
     {
         //std::cout<<r.getPoint(tt).y()<<std::endl;
         if(t.isIn(r.getPoint(tt))){
@@ -234,8 +234,8 @@ bool Ray::intersectAdvanced(Terrain &t, Vector2D& a, Vector2D& b, Vector3D& resu
 {
 
     /*resu=origine;
-    float min =t.getHauteurMin(a,b);
-    float max =t.getHauteurMax(a,b);
+    double min =t.getHauteurMin(a,b);
+    double max =t.getHauteurMax(a,b);
 
     Vector3D aBox(a.x(), min, a.y());
     Vector3D bBox(b.x(), max, b.y());
@@ -248,11 +248,11 @@ bool Ray::intersectAdvanced(Terrain &t, Vector2D& a, Vector2D& b, Vector3D& resu
         //std::cout<<"return false"<<std::endl;
         return false;
     }
-    float distInOut = in.distanceToPoint(out);
+    double distInOut = in.distanceToPoint(out);
     Vector3D dir=direction;
     Ray r(in,dir);
-    float penteMax=t.getPenteMax(a,b);
-    for(float tt = 0; tt <= distInOut+pas; tt+= pas)
+    double penteMax=t.getPenteMax(a,b);
+    for(double tt = 0; tt <= distInOut+pas; tt+= pas)
     {
         //std::cout<<r.getPoint(tt).y()<<std::endl;
         Vector3D point=r.getPoint(tt);
@@ -262,8 +262,8 @@ bool Ray::intersectAdvanced(Terrain &t, Vector2D& a, Vector2D& b, Vector3D& resu
         }
 
         Vector2D p(point.x(),point.z());
-        float hauteur=t.getHauteur(p);
-        float dif=point.y()-hauteur;
+        double hauteur=t.getHauteur(p);
+        double dif=point.y()-hauteur;
         if (direction.y()>penteMax){
             return false;
         }
@@ -275,4 +275,4 @@ bool Ray::intersectAdvanced(Terrain &t, Vector2D& a, Vector2D& b, Vector3D& resu
     return false;
 }
 
-float Ray::pas=0.1;
+double Ray::pas=0.1;
