+++ Synthese/terrain.h
@@ -8,7 +8,7 @@ class Terrain
 {
 
 public:
-    float getHauteurNormale(const Vector2D& p, Vector3D &n);
+    double getHauteurNormale(const Vector2D& p, Vector3D &n);
 
 protected:
     static double epsilon;
@@ -21,9 +21,9 @@ public:
     Terrain(const Vector2D& aa, const Vector2D& bb);
     bool isIn(const Vector3D& p);
     bool isOut(const Vector3D& p);
-    virtual float getHauteur(const Vector2D& p)=0;
-    virtual float getHauteurMin(Vector2D a,Vector2D b)=0;
-    virtual float getHauteurMax(Vector2D a,Vector2D b)=0;
+    virtual double getHauteur(const Vector2D& p)=0;
+    virtual double getHauteurMin(Vector2D a,Vector2D b)=0;
+    virtual double getHauteurMax(Vector2D a,Vector2D b)=0;
     virtual double getPenteMax(Vector2D a,Vector2D b)=0;
 };
