+++ Synthese/terrainoctaves.cpp
@@ -5,9 +5,9 @@ TerrainOctaves::TerrainOctaves(QVector<Octave> &o, const Vector2D& aa, const Vec
 {
 }
 
-float TerrainOctaves::getHauteur(const Vector2D &p)
+double TerrainOctaves::getHauteur(const Vector2D &p)
 {
-    float resu=0;
+    double resu=0;
     int length= octaves.length();
     for(int i = 0; i <length; ++i)
     {
@@ -17,11 +17,11 @@ float TerrainOctaves::getHauteur(const Vector2D &p)
     return resu;
 }
 
-float TerrainOctaves::getHauteurMin(Vector2D a, Vector2D b)
+double TerrainOctaves::getHauteurMin(Vector2D a, Vector2D b)
 {
-    float min=getHauteur(b);
+    double min=getHauteur(b);
     Vector2D vtmp;
-    float tmp;
+    double tmp;
     for(int i=(int)a.y();i<b.y();i++){
         for(int j=(int)a.x();j<b.x();j++){
             vtmp=Vector2D(j,i);
@@ -33,11 +33,11 @@ float TerrainOctaves::getHauteurMin(Vector2D a, Vector2D b)
     return min;
 }
 
-float TerrainOctaves::getHauteurMax(Vector2D a,Vector2D b)
+double TerrainOctaves::getHauteurMax(Vector2D a,Vector2D b)
 {
-    float max=getHauteur(b);
+    double max=getHauteur(b);
     Vector2D vtmp;
-    float tmp;
+    double tmp;
     for(int i=(int)a.y();i<b.y();i++){
         for(int j=(int)a.x();j<b.x();j++){
             vtmp=Vector2D(j,i);
@@ -52,11 +52,11 @@ float TerrainOctaves::getHauteurMax(Vector2D a,Vector2D b)
 double TerrainOctaves::getPenteMax(Vector2D a,Vector2D b)
 {
 
-    float max=0;
+    double max=0;
     Vector2D vtmp;
     Vector2D vtmp2;
     int tmp;
-    float htmp;
+    double htmp;
     for(int i=(int)a.y();i<b.y();i++){
         for(int j=(int)a.x();j<b.x();j++){
             vtmp=Vector2D(j,i);
