+++ Synthese/terrainpentex.h
@@ -7,9 +7,9 @@ class TerrainPenteX:public Terrain
 {
 public:
     TerrainPenteX(Vector2D& a, Vector2D& b);
-    virtual float getHauteur(const Vector2D &p);
-    virtual float getHauteurMin(Vector2D a,Vector2D b);
-    virtual float getHauteurMax(Vector2D a,Vector2D b);
+    virtual double getHauteur(const Vector2D &p);
+    virtual double getHauteurMin(Vector2D a,Vector2D b);
+    virtual double getHauteurMax(Vector2D a,Vector2D b);
     virtual double getPenteMax(Vector2D a,Vector2D b);
 };
